pull out to .h file shared across overlays
